requirements.txt
requests~=2.32.3
colorama~=0.4.6
python-dotenv~=1.0.1
pytz~=2024.1
pillow~=10.4.0
jira~=3.8.0
beautifulsoup4~=4.12.3